backend/requirements.txt
apscheduler
bcrypt
fastapi
jpholiday
passlib
pydantic
PyJWT
pytz
ruff
ruff-lsp
sqlalchemy
uvicorn
